backend/voice_tester/requirements.txt
# Voice Tester Requirements
# Requires Python 3.12+

# AWS SDK
boto3>=1.42.0
botocore>=1.42.0

# Nova Sonic SDK (speech-to-speech)
smithy-core>=0.3.0
smithy-aws-core>=0.4.0
smithy-http>=0.3.0
aws-sdk-bedrock-runtime>=0.4.0

# CDK
aws-cdk-lib>=2.100.0
constructs>=10.0.0

# CLI and utilities
PyYAML>=6.0
python-dotenv>=1.0.0
click>=8.0.0

# WebRTC support
websockets>=11.0.0
aiortc>=1.5.0
aiohttp>=3.9.0

# Audio 
sounddevice>=0.4.0
numpy>=1.24.0

# Reactive extensions
rx>=3.2.0

# Optional: for local testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
moto>=4.0.0
